Don't delete the output dirs in prepare
In some builds, this script is invoked several times in a single
target. The removal of the dist dirs leads to the output of the
first invocation to be deleted at the beginning of the second
invocation.

Test: atest mainline_modules_sdks_test
Change-Id: I39620b849ae8029719dc6cf150dae5092a093ea9
(cherry picked from commit d6a3166fe594e5900d7e19ef3a12e9751a409919)
Merged-In: I39620b849ae8029719dc6cf150dae5092a093ea9