Attempt to fix merge problems with min_sdk_version and 3p Rust crates.
This is causing auto-merger problems, likely as a result of ag/30694868
and possibly others.

This change was produced by diffing this file against the version in AOSP
and deleting any lines for Rust crates that are duplicate specifications
with minSdkVersion:34

Test: treehugger
Change-Id: Ib4d6f1657f3e1f8ecb81866cd27cdb706e0fa497
Aosp-First: No, because the version in internal is the one that seems to be incorrect and is causing problems.
Apex-Size-Increase: N/A
Previous-Platform-Support: N/A
Test-Info: Treehugger
Flag: EXEMPT: bug fix.