Fix allowed_deps.txt to list files in alphabetical order
This is caused by some hidden bug which has been magically fixed by the
change in this topic:
https://android-review.git.corp.google.com/q/topic:%22non-aosp-allowed-deps%22.
I have to merge this change first in order to land the AOSP changes
otherwise the Presubmit complains.

The change was created by running `(croot &&
packages/modules/common/build/update-apex-allowed-deps.sh)` on top of
the AOSP changes.

Ignore-AOSP-First: the targets in change are not in AOSP (they were
introduced in ag/30406176)

Bug: 380222284

Test: m
Change-Id: I1bd5647ecbf5f8ec263b31b490aa2eef1e4af61b